+++ API/src/services/recentsearch.service.ts
@@ -3,10 +3,15 @@ import recentsearchRepo from "../repos/recentsearch.repo";
 import { BadRequestError, UnauthorizedError } from "../core/error.response";
 import userRepo from "../repos/user.repo";
 import { convertStringToObjectId } from "../utils";
+import getMessageError from "../helpers/getMessageError";
 
 class RecentSearchService {
   constructor() {}
   async findRecentSearchByUser(userId: Types.ObjectId) {
+    const user = await userRepo.findById(userId);
+    if (!user) {
+      throw new UnauthorizedError(getMessageError(101));
+    }
     let recentSearch = await recentsearchRepo.findRecentSearchByUser(userId);
     if (!recentSearch) {
       recentSearch = await recentsearchRepo.createRecentSearch(userId);
@@ -20,14 +25,18 @@ class RecentSearchService {
     userId: Types.ObjectId,
     searchedUserId: string
   ) {
+    const user = await userRepo.findById(userId);
+    if (!user) {
+      throw new UnauthorizedError(getMessageError(101));
+    }
     if (!isValidObjectId(searchedUserId)) {
-      throw new UnauthorizedError("Invalid searched user id");
+      throw new UnauthorizedError(getMessageError(116));
     }
     const searchedUser = await userRepo.findById(
       convertStringToObjectId(searchedUserId)
     );
     if (!searchedUser) {
-      throw new BadRequestError(`User with id ${searchedUserId} is not found!`);
+      throw new BadRequestError(getMessageError(117));
     }
 
     const recentSearch =
@@ -43,14 +52,18 @@ class RecentSearchService {
     userId: Types.ObjectId,
     searchedUserId: string
   ) {
+    const user = await userRepo.findById(userId);
+    if (!user) {
+      throw new UnauthorizedError(getMessageError(101));
+    }
     if (!isValidObjectId(searchedUserId)) {
-      throw new UnauthorizedError("Invalid searched user id");
+      throw new UnauthorizedError(getMessageError(116));
     }
     const searchedUser = await userRepo.findById(
       convertStringToObjectId(searchedUserId)
     );
     if (!searchedUser) {
-      throw new BadRequestError(`User with id ${searchedUserId} is not found!`);
+      throw new BadRequestError(getMessageError(117));
     }
     const recentSearch = await recentsearchRepo.addSearchedUserToRecentSearch(
       userId,
@@ -61,6 +74,10 @@ class RecentSearchService {
     };
   }
   async removeAllSearchedUsersFromRecentSearch(userId: Types.ObjectId) {
+    const user = await userRepo.findById(userId);
+    if (!user) {
+      throw new UnauthorizedError(getMessageError(101));
+    }
     const recentSearch =
       await recentsearchRepo.removeAllSearchedUsersFromRecentSearch(userId);
     return {
